chore(website): break footer into separate files
