+++ packages/paste-website/src/components/site-wrapper-new/site-footer/SiteFooterIllustration.tsx
@@ -1,0 +1,15 @@
+import * as React from 'react';
+import {useTheme} from '@twilio-paste/core/theme';
+import {Box} from '@twilio-paste/core/box';
+import FooterBuildingBlocks from '../../../assets/illustrations/FooterBuildingBlocks.svg';
+
+const SiteFooterIllustration: React.FC<{}> = () => {
+  const theme = useTheme();
+  return (
+    <Box bottom={-6} display={['none', 'none', 'none', 'block']} right={theme.sizes.sizeSquare120} position="absolute">
+      <FooterBuildingBlocks aria-hidden="true" role="img" />
+    </Box>
+  );
+};
+
+export {SiteFooterIllustration};
